Implement node replacement and vector-based perp dist
Implement a dedicated function for node replacement, as well as changing the perpendicular distance function to use vectors, and take the segment rather than projected line into account.

Also add an additional "alias" attribute to act as a name for a GameObject, make the position attribute of GameObject private, and implement the edge-splitting function.